fix(f2.4): taban cizgisi yayincisinda %d/long int bicim uyarisini gider
Uyari: kitti_baseline_publisher_main.cpp:113:17: format '%d' expects
argument of type 'int', but argument 5 has type 'long int' [-Wformat=]

Sebep: `stride` `declare_parameter<int>` ile alinir ama rclcpp tamsayi
parametreyi `int64_t` dondurur ve `const auto` onu yakalar. `%d` ile basmak
tanimsiz davranistir.

Duzeltme: bicim `%ld`, arguman acikca `long`'a cevrildi. Daraltip `%d`
birakmak da uyariyi susturur ama buyuk bir parametre degerini LOG'da sessizce
kirpardi; deger oldugu gibi yazilir. Plan alanlari `int`'tir, `%d`'leri
dogrudur ve dokunulmadi. DAVRANIS DEGISMEDI: yalniz bicim dizesi ve bir cast.

KOSTURULAN: temiz colcon build (kerteriz_bringup artik stderr URETMIYOR),
colcon test 118 test 0 hata, pre-commit (clang-format dahil) temiz.
Baska temizlik YAPILMADI.

+++ kerteriz_bringup/src/kitti_baseline_publisher_main.cpp
@@ -109,11 +109,18 @@ int main(int argc, char** argv) {
     // ozet tam da yakalamasi beklenen ayrismayi gizlerdi. Iki sureci
     // karsilastirmak icin stride + olcum sayisi + ozet uclusune bakilir.
     const std::uint64_t ozet = kerteriz_bringup::measurement_stamp_digest(plan);
+    // `stride` `declare_parameter<int>`'ten gelir ama rclcpp tamsayi
+    // parametreyi `int64_t` olarak dondurur; `auto` da onu yakalar. Bu yuzden
+    // bicim `%ld` ve arguman acikca `long`'a cevrilir — `%d` ile basmak
+    // tanimsiz davranisti. Daraltip `%d` birakmak da olurdu ama o, buyuk bir
+    // parametre degerini LOG'da sessizce kirpardi; burada deger oldugu gibi
+    // yazilir. Plan alanlari `int`'tir, onlarin `%d`'si dogrudur.
     RCLCPP_INFO(node->get_logger(),
-                "GNSS konum seyreltmesi acik: stride=%d aday=%d secili_slot=%d olcum=%d "
+                "GNSS konum seyreltmesi acik: stride=%ld aday=%d secili_slot=%d olcum=%d "
                 "ara-degerlenmis-atlanan=%d damga_ozeti=%llu",
-                stride, plan.candidate_count, plan.selected_slot_count, plan.selected_usable_count,
-                plan.selected_interpolated_skipped, static_cast<unsigned long long>(ozet));
+                static_cast<long>(stride), plan.candidate_count, plan.selected_slot_count,
+                plan.selected_usable_count, plan.selected_interpolated_skipped,
+                static_cast<unsigned long long>(ozet));
   }
 
   auto saat = node->create_publisher<rosgraph_msgs::msg::Clock>("/clock", 10);
